Refactor: search component
